+++ Rapa/RapaGame/RapaduraEngine/Entities/PreBuilt/Solids/Tile.cs
@@ -7,7 +7,7 @@ public class Tile : Solid
 {
     #region constructor
     
-    public Tile(Texture2D texture, float width = 8, float height = 8, bool debugMode = false) : base(texture, width, height, debugMode)
+    public Tile(Texture2D texture, int width = 8, int height = 8, bool debugMode = false) : base(texture, width, height, debugMode)
     {
     }
